src/main/java/net/silthus/chat/listeners/PlayerListener.java
package net.silthus.chat.listeners;

import net.silthus.chat.SChat;
import org.bukkit.event.EventHandler;
import org.bukkit.event.Listener;
import org.bukkit.event.player.PlayerJoinEvent;
import org.bukkit.event.player.PlayerQuitEvent;

public class PlayerListener implements Listener {

    private final SChat plugin;

    public PlayerListener(SChat plugin) {
        this.plugin = plugin;
    }

    @EventHandler(ignoreCancelled = true)
    public void onJoin(PlayerJoinEvent event) {
        plugin.getChatManager().registerChatter(event.getPlayer());
    }

    @EventHandler(ignoreCancelled = true)
    public void onQuit(PlayerQuitEvent event) {
        plugin.getChatManager().unregisterChatter(event.getPlayer());
    }
}
